requirements.txt
discord.py==2.5.2
yt-dlp==2025.5.22
PyNaCl==1.5.0